- apply suggestions
