+++ tests/unit/resources/config/opensearch-security/internal_users.yml
@@ -9,8 +9,12 @@ _meta:
 
 ## Demo users
 
+kibanaserver:
+  hash: $2b$12$VNfTvrDDZ5JjT9.BZDqRTOh6sZoUxdMmgLueB4X0H93E0/e9FinRC
+  reserved: false
+  description: Kibanaserver user
 admin:
-  hash: $2b$12$Cts2k6vk.emW.E1mQN.B5u/5DfwKybaXFRREHfIwUs7RUHNp8h.c.
+  hash: $2b$12$3wL/WsFm7y0S2jyfiyHV4elgGEg2jeXrPFUmYdZGkj5aDF9hpCPFi
   reserved: false
   backend_roles:
   - admin
@@ -18,7 +22,3 @@ admin:
   - security_rest_api_access
   - all_access
   description: Admin user
-kibanaserver:
-  hash: $2b$12$0ZYW/CiICy1.7YXpRYDjeOLD75qnPAu4I3w5KXZmJvFp1rRiCm8qa
-  reserved: false
-  description: Kibanaserver user
